requirements.txt
google-generativeai
psycopg2-binary
python-dotenv
streamlit
pandas
sqlalchemy
langchain
langchain-google-genai
langgraph
